+++ README.md
@@ -1,11 +1,10 @@
-# ACHO
+# ConfOpt
 
 [![License](https://img.shields.io/badge/License-Apache_2.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
 [![arXiv](https://img.shields.io/badge/arXiv-ACHO-cyan)](https://doi.org/10.48550/arXiv.2207.03017)
 
-ACHO (Adaptive Conformal Hyperparameter Optimization) is
-an inferential hyperparameter optimization package designed to
+ConfOpt is an inferential hyperparameter optimization package designed to
 speed up model hyperparameter tuning.
 
-The original methods and preprint for ACHO
-can be found [here](https://doi.org/10.48550/arXiv.2207.03017).
+The package currently implements Adaptive Conformal Hyperparameter Optimization (ACHO), as detailed
+in [the original paper](https://doi.org/10.48550/arXiv.2207.03017).
