+++ examples/tabular_tuning.py
@@ -1,7 +1,7 @@
 from sklearn.datasets import fetch_california_housing
 from sklearn.ensemble import RandomForestRegressor
 
-from acho.tuning import ConformalSearcher
+from confopt.tuning import ConformalSearcher
 
 # Set up toy data:
 X, y = fetch_california_housing(return_X_y=True)
